+++ python/fate_arch/federation/__init__.py
@@ -1,4 +1,11 @@
-from fate_arch.federation._type import FederationEngine
 from fate_arch.federation._split import segment_transfer_enabled
+from fate_arch.federation._type import FederationEngine
+from fate_arch.federation._federation import Tag, get, remote
 
-__all__ = ['FederationEngine', 'segment_transfer_enabled']
+__all__ = [
+    "FederationEngine",
+    "segment_transfer_enabled",
+    "Tag",
+    "remote",
+    "get",
+]
